move portal into tooltip, reduces stale state
