+++ packages/design-system/src/components/tooltip/index.js
@@ -26,6 +26,9 @@ import {
   useCallback,
   useDebouncedCallback,
   useEffect,
+  useLayoutEffect,
+  useResizeEffect,
+  createPortal,
 } from '@googleforcreators/react';
 
 /**
@@ -35,9 +38,10 @@ import { PLACEMENT } from '../popup';
 import { prettifyShortcut } from '../keyboard';
 import { THEME_CONSTANTS } from '../../theme';
 import { Text } from '../typography';
-import { RTL_PLACEMENT } from '../popup/constants';
+import { RTL_PLACEMENT, PopupContainer } from '../popup/constants';
+import { getOffset, getTransforms } from '../popup/utils';
 import { noop } from '../../utils';
-import TooltipPopup from './tooltipPopup';
+// import TooltipPopup from './tooltipPopup';
 import { SvgForTail, Tail, SVG_TOOLTIP_TAIL_ID, TAIL_HEIGHT } from './tail';
 
 const SPACE_BETWEEN_TOOLTIP_AND_ELEMENT = TAIL_HEIGHT;
@@ -134,6 +138,11 @@ function Tooltip({
   const [dynamicPlacement, setDynamicPlacement] = useState(placement);
   const isMounted = useRef(false);
 
+  const [popupState, setPopupState] = useState(null);
+  const isPopupMounted = useRef(false);
+  const popup = useRef(null);
+  const isOpen = Boolean(shown && (shortcut || title));
+
   const spacing = useMemo(
     () => ({
       x:
@@ -147,6 +156,26 @@ function Tooltip({
     }),
     [placement]
   );
+
+  const positionPopup = useCallback(() => {
+    if (!isPopupMounted.current || !anchorRef?.current) {
+      return;
+    }
+
+    setPopupState({
+      offset: anchorRef.current
+        ? getOffset({
+            placement,
+            spacing,
+            anchor: forceAnchorRef || anchorRef,
+            popup,
+            isRTL,
+            ignoreMaxOffsetY,
+            leftOffset,
+          })
+        : {},
+    });
+  }, [placement, spacing, isRTL, ignoreMaxOffsetY, leftOffset, forceAnchorRef]);
 
   // When near the edge of the viewport we want to force the tooltip to a new placement as to not
   // cutoff the contents of the tooltip.
@@ -197,7 +226,6 @@ function Tooltip({
   const resetPlacement = useDebouncedCallback(() => {
     setDynamicPlacement(placementRef.current);
   }, 100);
-
   const delay = useRef();
   const onHover = useCallback(
     (evt) => {
@@ -248,6 +276,30 @@ function Tooltip({
     };
   }, []);
 
+  useLayoutEffect(() => {
+    if (!isOpen) {
+      return undefined;
+    }
+    isPopupMounted.current = true;
+
+    positionPopup();
+    // Adjust the position when scrolling.
+    document.addEventListener('scroll', positionPopup, true);
+    return () => {
+      document.removeEventListener('scroll', positionPopup, true);
+      isPopupMounted.current = false;
+    };
+  }, [isOpen, positionPopup]);
+
+  useLayoutEffect(() => {
+    if (!isPopupMounted.current) {
+      return;
+    }
+    positionArrow(popupState);
+  }, [popupState, positionArrow]);
+
+  useResizeEffect({ current: document.body }, positionPopup, [positionPopup]);
+
   return (
     <>
       <Wrapper
@@ -268,42 +320,49 @@ function Tooltip({
         {children}
       </Wrapper>
 
-      <TooltipPopup
-        anchor={forceAnchorRef || anchorRef}
-        placement={dynamicPlacement}
-        spacing={spacing}
-        isOpen={Boolean(shown && (shortcut || title))}
-        onPositionUpdate={positionArrow}
-        zIndex={popupZIndexOverride}
-        noOverFlow
-        ignoreMaxOffsetY={ignoreMaxOffsetY}
-        leftOffset={leftOffset}
-        isRTL={isRTL}
-      >
-        <TooltipContainer
-          className={className}
-          ref={tooltipRef}
-          placement={dynamicPlacement}
-          shown={shown}
-        >
-          <TooltipText size={THEME_CONSTANTS.TYPOGRAPHY.PRESET_SIZES.X_SMALL}>
-            {shortcut ? `${title} (${prettifyShortcut(shortcut)})` : title}
-          </TooltipText>
-          {hasTail && (
-            <>
-              <SvgForTail>
-                <clipPath
-                  id={SVG_TOOLTIP_TAIL_ID}
-                  clipPathUnits="objectBoundingBox"
+      {popupState && isOpen
+        ? createPortal(
+            <PopupContainer
+              ref={popup}
+              $offset={popupState.offset}
+              noOverFlow
+              zIndex={999999999999}
+              transforms={getTransforms(placement, isRTL)}
+            >
+              <TooltipContainer
+                className={className}
+                ref={tooltipRef}
+                placement={dynamicPlacement}
+                shown={shown}
+              >
+                <TooltipText
+                  size={THEME_CONSTANTS.TYPOGRAPHY.PRESET_SIZES.X_SMALL}
                 >
-                  <path d="M1,1 L0.868,1 C0.792,1,0.72,0.853,0.676,0.606 L0.585,0.098 C0.562,-0.033,0.513,-0.033,0.489,0.098 L0.399,0.606 C0.355,0.853,0.283,1,0.207,1 L0,1 L1,1" />
-                </clipPath>
-              </SvgForTail>
-              <Tail placement={dynamicPlacement} translateX={arrowDelta} />
-            </>
-          )}
-        </TooltipContainer>
-      </TooltipPopup>
+                  {shortcut
+                    ? `${title} (${prettifyShortcut(shortcut)})`
+                    : title}
+                </TooltipText>
+                {hasTail && (
+                  <>
+                    <SvgForTail>
+                      <clipPath
+                        id={SVG_TOOLTIP_TAIL_ID}
+                        clipPathUnits="objectBoundingBox"
+                      >
+                        <path d="M1,1 L0.868,1 C0.792,1,0.72,0.853,0.676,0.606 L0.585,0.098 C0.562,-0.033,0.513,-0.033,0.489,0.098 L0.399,0.606 C0.355,0.853,0.283,1,0.207,1 L0,1 L1,1" />
+                      </clipPath>
+                    </SvgForTail>
+                    <Tail
+                      placement={dynamicPlacement}
+                      translateX={arrowDelta}
+                    />
+                  </>
+                )}
+              </TooltipContainer>
+            </PopupContainer>,
+            document.body
+          )
+        : null}
     </>
   );
 }
